refactor: split oversized types.ts, settingsStore, and useHotExitRestore (partial #96)
- bridge/types.ts (870 lines) -> core-types.ts, document-types.ts,
  protocol-types.ts, workspace-types.ts with barrel re-export
- settingsStore.ts (521 lines) -> extract types to settingsTypes.ts,
  deepMerge to src/utils/deepMerge.ts
- useHotExitRestore.ts (497 lines) -> extract helpers to restoreHelpers.ts
