+++ src/utils/deepMerge.ts
@@ -1,0 +1,28 @@
+/**
+ * Deep merge utility for settings migration.
+ * Merges persisted state into current defaults, preserving new default properties.
+ */
+export function deepMerge<T extends Record<string, unknown>>(target: T, source: Partial<T>): T {
+  const result = { ...target };
+  for (const key of Object.keys(source) as Array<keyof T>) {
+    const sourceValue = source[key];
+    const targetValue = target[key];
+    if (
+      sourceValue !== null &&
+      typeof sourceValue === "object" &&
+      !Array.isArray(sourceValue) &&
+      targetValue !== null &&
+      typeof targetValue === "object" &&
+      !Array.isArray(targetValue)
+    ) {
+      result[key] = deepMerge(
+        targetValue as Record<string, unknown>,
+        sourceValue as Record<string, unknown>
+      ) as T[keyof T];
+    } else if (sourceValue !== undefined && sourceValue !== null) {
+      // Skip null values to preserve defaults (null from corrupted localStorage)
+      result[key] = sourceValue as T[keyof T];
+    }
+  }
+  return result;
+}
